- Add documentation. - Clean up unused code.
