+++ Mlem/Views/Shared/Cached Image.swift
@@ -14,11 +14,10 @@ import Nuke
 
 struct CachedImage: View {
     
-    @EnvironmentObject var quickLookPreview: QuickLookPreview
+    @EnvironmentObject var quickLookPaths: QuickLookPaths
     
     let url: URL?
     let shouldExpand: Bool
-    @State var bigPicMode: URL?
     
     // state vars to track the current image size and whether that size needs to be recomputed when the image actually loads. Combined with the image size cache, this produces good scrolling behavior except in the case where we scroll past an image and it derenders before it ever gets a chance to load, in which case that image will cause a slight hiccup on the way back up. That's kind of an unsolvable problem, since we can't know the size before we load the image at all, but that's fine because it shouldn't really happen during normal use. If we really want to guarantee smooth feed scrolling we can squish any image with no cached size into a square, but that feels like squishing a lot of images for the sake of a fringe case.
     @State var size: CGSize
@@ -110,23 +109,16 @@ struct CachedImage: View {
                                     }
                                     try data.write(to: quicklook)
                                     await MainActor.run {
-//                                        bigPicMode = quicklook
-                                        quickLookPreview.url = quicklook
+                                        quickLookPaths.url = quicklook
                                     }
                                 } catch {
                                     print(String(describing: error))
                                 }
                             }
                         }
-                        .fullScreenCover(item: $quickLookPreview.url) { url in
+                        .fullScreenCover(item: $quickLookPaths.url) { url in
                             QuickLookView(urls: [url])
                         }
-//                        .quickLookPreview($bigPicMode)
-//                        .onChange(of: bigPicMode) { mode in
-//                            if mode == nil, let dismissCallback {
-//                                dismissCallback()
-//                            }
-//                        }
                 } else {
                     imageView
                 }
